Project 2!!!
I know it's all still in the project 1 file, but the second project is
just adding onto the first. I've added a parser which checks the syntax
of the tokens provided by the scanner. It then inserts the contents of
each token into containers for schemes, facts, rules, and queries for
future datalog interpreting. The program works on at least two test
input files. More edits may be made if testing shows it does not work.