Tested Project2
After testing against most of the test input files, I believe the
project is finished. I needed to edit the expression() function so that
when there was an expression parameter it would insert itself into it's
parent predicate correctly. There may have been a better way, but I
thought the easiest way would be to make a new exParameter() function
that returned a Parameter object to the expression() function and didn't
add a parameter to the rule or query list.