Completely Finished!
I have turned in the project and it has passed all tests. I had to
remove my multi-line comment of the grammar because it caused a
warning... I also needed to edit my expression() descent. If there were
nested expressions, it would add the deeper expressions as parameters
twice, once on their own and once in the larger expression.