Build the Rules Evaluation
Rules are now evaluated and add tuples to relations in the database on
creation. This process reused some of the code that was used with the
query evaluation, so I simplified this code and broke it into many
smaller functions. Now, for the assignment, I need to print out the
results of the rule evaluation as specified.